+++ gemrb/core/SymbolMgr.h
@@ -47,7 +47,7 @@ public:
 	virtual int GetValueIndex(unsigned int Index) const = 0;
 	virtual int FindValue(int val) const = 0;
 	virtual int FindString(char *str, int len) const = 0;
-	virtual int GetSize() const = 0;
+	virtual size_t GetSize() const = 0;
 	virtual int GetHighestValue() const = 0;
 };
